Routes + Features

+++ src/main/java/model/BaseEntity.java
@@ -1,5 +1,5 @@
 package model;
 
 public interface BaseEntity {
-  public long getId();
+  public int getId();
 }
